// apps/website/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
      tool_capabilities: {
        Row: {
          created_at: string
          description: string
          id: number
          key: string
          name: string
          tool_id: number | null
        }
        Insert: {
          created_at?: string
          description: string
          id?: number
          key: string
          name: string
          tool_id?: number | null
        }
        Update: {
          created_at?: string
          description?: string
          id?: number
          key?: string
          name?: string
          tool_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "atm_tool_capabilities_tool_id_fkey"
            columns: ["tool_id"]
            referencedRelation: "tools"
            referencedColumns: ["id"]
          },
        ]
      }
      tools: {
        Row: {
          category: Database["public"]["Enums"]["tool_categories"] | null
          created_at: string
          description: string | null
          id: number
          name: string
          owner_id: string
          owner_username: string
          tool_handle: string
          type: string
        }
        Insert: {
          category?: Database["public"]["Enums"]["tool_categories"] | null
          created_at?: string
          description?: string | null
          id?: number
          name: string
          owner_id: string
          owner_username: string
          tool_handle: string
          type: string
        }
        Update: {
          category?: Database["public"]["Enums"]["tool_categories"] | null
          created_at?: string
          description?: string | null
          id?: number
          name?: string
          owner_id?: string
          owner_username?: string
          tool_handle?: string
          type?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      tool_categories:
        | "Uncategorized"
        | "Communication"
        | "Productivity"
        | "Collaboration"
        | "Social Media"
        | "Analytics"
        | "Finance"
        | "Fun"
        | "Utility"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
    ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never